fix creation of .xor file
commit d12d0803e625e6c0d5cef2be8b91009e8041eaee, apart from introducing
a "Broken SKA" check, which doesn't quite work, also increased size of
the buffers involved in the shared key auth, but out of 3 occurences which
used the hardcoded buffer size for memset, only one was changed!
this caused the buffer to contain garbage, and write an invalid xor file.

i now use sizeof(), which always works correctly with any buffersize...
this should have been done this way from the beginning.
with this change, and the commented out "Broken SKA" check, i have been
able to produce the same working .xor file from the same pcap as the
last known good version, aircrack-ng-1.0-beta2, and could successfully
use it to auth to the test AP using WEP with SKA.