prevent "Broken SKA" check from not writing potentially correct xor files
this check is often triggered even with perfectly good WEP SKA captures,
and caused an early exit in the past without writing a .xor file at all.
due to this, since it was introduced 9 years ago, many people have been
unable to successfully create a .xor file; hereby heavily crippling
the available WEP attack toolkit.

we now check whether the .xor file is already existing, and when the
broken SKA check has been hit, the file will not be overwritten, but
if none is existing yet, it will be created anyways.

the commit introducing the check was d12d0803e625e6c0d5cef2be8b91009e8041eaee .

it not only broke the SKA detection in many cases, it even broke the
.xor file generation itself, since it increased several buffer's size
without properly adjusting the related memset calls on those buffers.
we're fixing that issue in a separate commit.

since the bad commit broke 2 things, it may even be possible that all
the involved buffers could be reverted to their original size, but
i have not investigated that.

closes #125